Interactive MCA variables plot

This function generates an HTML widget displaying the variables plot of an MCA result.

MCA_var_plot(
  res,
  xax = 1,
  yax = 2,
  var_sup = TRUE,
  var_sup_choice = NULL,
  var_lab_min_contrib = 0,
  point_size = 64,
  labels_prepend_var = FALSE,
  col_var = NULL,
  symbol_var = NULL,
  size_var = NULL,
  size_range = c(10, 300),
  zoom_callback = NULL,
  in_explor = FALSE,
  ...
)

Arguments

res
Result of prepare_results() call

xax
Horizontal axis number

yax
Vertical axis number

var_sup
TRUE to display supplementary variables

var_sup_choice
list of supplementary variables to display

var_lab_min_contrib
Contribution threshold to display points labels

point_size
base point size

labels_prepend_var
if TRUE, prepend variable names to labels

col_var
name of the variable for points color

symbol_var
name of the variable for points symbol

size_var
name of the variable for points size

size_range
points size range with format c(minimum, maximum)

zoom_callback
scatterD3 zoom callback JavaScript body

in_explor
wether the plot is to be displayed in the explor interface

...
Other arguments passed to scatterD3