Interactive PCA indivuals plot

This function generates an HTML widget displaying the individuals plot of a PCA result.

PCA_ind_plot(
  res,
  xax = 1,
  yax = 2,
  ind_sup = TRUE,
  ind_lab_min_contrib = 0,
  col_var = NULL,
  symbol_var = NULL,
  opacity_var = NULL,
  size_var = NULL,
  size_range = c(10, 300),
  lab_var = NULL,
  zoom_callback = NULL,
  in_explor = FALSE,
  ...
)

Arguments

res

Result of prepare_results() call

xax

Horizontal axis number

yax

Vertical axis number

ind_sup

TRUE to display supplementary individuals

ind_lab_min_contrib

Contribution threshold to display points labels

col_var

variable to be used for points color

symbol_var

name of the variable for points symbol

opacity_var

name of the variable for points opacity

size_var

name of the variable for points size

size_range

points size range with format c(minimum, maximum)

lab_var

variable to be used for points names

zoom_callback

scatterD3 zoom callback JavaScript body

in_explor

wether the plot is to be displayed in the explor interface

...

Other arguments passed to scatterD3