Interactive PCA variables plot

This function generates an HTML widget displaying the variables plot of a PCA result.

PCA_var_plot(
  res,
  xax = 1,
  yax = 2,
  var_sup = TRUE,
  var_sup_choice = NULL,
  var_lab_min_contrib = 0,
  scale_unit = FALSE,
  col_var = NULL,
  size_var = NULL,
  zoom_callback = NULL,
  in_explor = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

Arguments

res
Result of prepare_results() call

xax
Horizontal axis number

yax
Vertical axis number

var_sup
TRUE to display supplementary variables

var_sup_choice
list of supplementary variables to display

var_lab_min_contrib
Contribution threshold to display points labels

scale_unit
wether the PCA is scaled

col_var
name of the variable for points color

size_var
name of the variable for points size

zoom_callback
scatterD3 zoom callback JavaScript body

in_explor
wether the plot is to be displayed in the explor interface

xlim
custom x axis limits

ylim
custom y axis limits

...
Other arguments passed to scatterD3