Compute supplementary variables data for a GDAtools::speMCA result

Compute supplementary variables data for a GDAtools::speMCA result

speMCA_varsup(mca, df)

Arguments

mca
result object from speMCA.

df
data frame with the supplementary variables data. Must have the same number of rows than the data used with speMCA.

Value

A list of results suitable to be added as a `supv` element to the `mca` object.

See also