Interface for analysis results exploration

This function launches a shiny app in a web browser in order to do interactive visualisation and exploration of an analysis results.

explor(obj)

# S3 method for CA
explor(obj)

# S3 method for coa
explor(obj)

# S3 method for MCA
explor(obj)

# S3 method for speMCA
explor(obj)

# S3 method for mca
explor(obj)

# S3 method for acm
explor(obj)

# S3 method for PCA
explor(obj)

# S3 method for princomp
explor(obj)

# S3 method for prcomp
explor(obj)

# S3 method for pca
explor(obj)

Arguments

obj
object containing analysis results

Value

The function launches a shiny app in the system web browser.

Details

If you want to display supplementary individuals or variables and you're using the dudi.coa function, you can add the coordinates of suprow and/or supcol to as supr and/or supr elements added to your dudi.coa result (See example).

If you want to display supplementary individuals or variables and you're using the dudi.acm function, you can add the coordinates of suprow and/or supcol to as supi and/or supv elements added to your dudi.acm result (See example).

If you want to display supplementary individuals or variables and you're using the dudi.pca function, you can add the coordinates of suprow and/or supcol to as supi and/or supv elements added to your dudi.pca result (See example).

Examples

# NOT RUN {
require(FactoMineR)

## FactoMineR::MCA exploration
data(hobbies)
mca <- MCA(hobbies[1:1000,c(1:8,21:23)], quali.sup = 9:10,
           quanti.sup = 11, ind.sup = 1:100, graph = FALSE)
explor(mca)

## FactoMineR::PCA exploration
data(decathlon)
d <- decathlon[,1:12]
pca <- PCA(d, quanti.sup = 11:12, graph = FALSE)
explor(pca)
# }# NOT RUN {
library(ade4)

data(bordeaux)
tab <- bordeaux
row_sup <- tab[5,-4]
col_sup <- tab[-5,4]
coa <- dudi.coa(tab[-5,-4], nf = 5, scannf = FALSE)
coa$supr <- suprow(coa, row_sup)$lisup
coa$supc <- supcol(coa, col_sup)$cosup
explor(coa)
# }# NOT RUN {
library(ade4)
data(banque)
d <- banque[-(1:100),-(19:21)]
ind_sup <- banque[1:100, -(19:21)]
var_sup <- banque[-(1:100),19:21]
acm <- dudi.acm(d, scannf = FALSE, nf = 5)
acm$supv <- supcol(acm, dudi.acm(var_sup, scannf = FALSE, nf = 5)$tab)$cosup
colw <- acm$cw*ncol(d)
X <- acm.disjonctif(ind_sup)
X <- data.frame(t(t(X)/colw) - 1)
acm$supi <- suprow(acm, X)$lisup
explor(acm)
# }# NOT RUN {
library(ade4)
data(deug)
d <- deug$tab
sup_var <- d[-(1:10), 8:9]
sup_ind <- d[1:10, -(8:9)]
pca <- dudi.pca(d[-(1:10), -(8:9)], scale = TRUE, scannf = FALSE, nf = 5)
supi <- suprow(pca, sup_ind)
pca$supi <- supi$lisup
supv <- supcol(pca, dudi.pca(sup_var, scale = TRUE, scannf = FALSE)$tab)
pca$supv <- supv$cosup
explor(pca)
# }